+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
-    * BUGFIX      #4562 [WebsiteBundle]         Fix BinaryFileResponse of html files.
+    * BUGFIX      #4540  [ContentBundle]        Fix copying of teaser selection in WebspaceCopyCommand
+    * BUGFIX      #4562  [WebsiteBundle]        Fix BinaryFileResponse of html files.
     * FEATURE     #4532  [MediaBundle]          Add command to remove format cache files which not longer exists
     * HOTFIX      #4512  [Components]           Fix pathcleanup whitespace character problems
     * FEATURE     #4504  [WebsiteBundle]        Dispatch an event after cache clear.